Pathnodes updates
Intersections now operational with give-way to other path nodes + related updates to car ai to make sure cars give way if we set up the rules to do so. Minor tweeks to pathnode progresstracker to make sure the car always looks to the front of the car and never behind. Also enabled multiple path node editing, and better performance to pathfinding search algoritm